+++ macos/script/package_app.sh
@@ -7,11 +7,11 @@ BUNDLE_ID="io.github.soleillevant0125.CodexTokenOverlay"
 MIN_SYSTEM_VERSION="14.0"
 ARCH="$(uname -m)"
 CONFIGURATION="release"
-VERSION="0.2.0"
+VERSION="0.2.1"
 OUTPUT_DIR=""
 
 usage() {
-  echo "usage: $0 [--arch arm64|x86_64] [--configuration debug|release] [--version 0.2.0] [--output directory]" >&2
+  echo "usage: $0 [--arch arm64|x86_64] [--configuration debug|release] [--version 0.2.1] [--output directory]" >&2
 }
 
 while [[ $# -gt 0 ]]; do